pyproject.toml
[project]
name = "catalog-of-complexity"
version = "0.1.0"
description = "Catalog of Complexity: comparative research operating system for complex systems."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Nicholas Underwood" }]
dependencies = [
  "jsonschema>=4.21",
  "ruamel.yaml>=0.18",
  "pydantic>=2.6",
  "duckdb>=1.0",
  "pyarrow>=15.0",
  "pandas>=2.2",
  "click>=8.1",
  "rich>=13.7",
  "rdflib>=7.0",
  "frictionless>=5.16",
  "rocrate>=0.11",
  "python-dateutil>=2.9",
  "python-ulid>=3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.3",
  "mypy>=1.8",
  "pre-commit>=3.7",
]

[project.scripts]
coc = "coc.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coc"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
